Refactor: Restructure Headers & Footers as LayoutWrapper
